Fixed compilation errors.
Fix compilation errors and type errors introduced by the previous
commit. This commit adds correct types to Signal.getPromise, fixes
some type casts, reorders some shared exports inside the Keyboard
folder and removes several unused imports, removing about 80 errors.

Since the ScreenKeyboard is not used at the moment, but the TypeScript
compiler still checks it for errors, it has temporarily been renamed to
have a .IGN extension.
