+++ src/Keyboard/index.ts
@@ -1,2 +1,3 @@
-export { Keyboard } from "./Keyboard";
+export * from "./types";
+export { Keyboard, KeyboardSource } from "./Keyboard";
 export { PhysicalKeyboard } from "./PhysicalKeyboard";
